Fixed error in is_prime(n) method

+++ python/isprime.py
@@ -11,7 +11,7 @@ def is_prime(n):
         return True
     if n % 2 == 0 or n % 3 == 0:
         return False
-    for i in range(5, math.sqrt(n) + 1, 6):
+    for i in range(5, int(math.sqrt(n) + 1), 6):
         if n % i == 0 or n % (i+2) == 0:
             return False
     return True
